+++ com.unity.render-pipelines.high-definition/HDRP/Material/Builtin/BuiltinData.cs
@@ -35,8 +35,10 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             [SurfaceDataAttributes("Shadow Mask 3")]
             public float shadowMask3;
 
-            [SurfaceDataAttributes("Emissive Color", false, false)]
+            [SurfaceDataAttributes("Emissive Color", false, true)]
             public Vector3 emissiveColor;
+            [SurfaceDataAttributes("Emissive Intensity")]
+            public float emissiveIntensity;
 
             // These is required for motion blur and temporalAA
             [SurfaceDataAttributes("Velocity")]